Cleanup (#341)
* directory tree copy and clean up

* clean up template comments

* clean up comments

* remove konsole-part from version lock

* simplify packages version lock

* move rpmfusion info

* clean up comments and syntax

* cleanup comments and syntax

* cleanup comments

* add package replacment check to focusrite packages

* comment syntax

* cleanup comments and syntax

* fix system_files copy

* remove .gitkeep files

+++ build_files/install-focusrite-scarlett.sh
@@ -2,13 +2,21 @@
 
 set -ouex pipefail
 
-### Packages array
-# Target directories as elements
+# Packages to install
 packages=(
     /ctx/repo_files/fcp-support/*
     /ctx/repo_files/scarlett4-firmware/*
     /ctx/repo_files/alsa-scarlett-gui/*
 )
 
-### Install packages
+# Check if base image packages are being replaced with a dry run
+dnf5 --setopt=tsflags=test -y install "${packages[@]}" 2>&1 | tee /tmp/dryrun.log
+
+# Check log for upgrading and downgrading
+if grep -qE '^(Upgrading|Downgrading):' /tmp/dryrun.log; then
+    echo "::notice::Detected package replacements. Aborting build."
+    exit 1
+fi
+
+# Install packages
 dnf5 -y install "${packages[@]}"
